Move FileLock cluster service to its own module
