+++ extensions/file-cluster-service/deployment/pom.xml
@@ -21,13 +21,13 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.camel.quarkus</groupId>
-        <artifactId>camel-quarkus-file-parent</artifactId>
+        <artifactId>camel-quarkus-file-cluster-service-parent</artifactId>
         <version>3.10.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>camel-quarkus-file-deployment</artifactId>
-    <name>Camel Quarkus :: File :: Deployment</name>
+    <artifactId>camel-quarkus-file-cluster-service-deployment</artifactId>
+    <name>Camel Quarkus :: File Cluster Service :: Deployment</name>
 
     <dependencies>
         <dependency>
@@ -36,7 +36,16 @@
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-file</artifactId>
+            <artifactId>camel-quarkus-file-deployment</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.camel.quarkus</groupId>
+            <artifactId>camel-quarkus-file-cluster-service</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-junit5-internal</artifactId>
+            <scope>test</scope>
         </dependency>
     </dependencies>
