+++ integration-tests/master-file/src/main/resources/application.properties
@@ -17,7 +17,6 @@
 
 application.id = leader
 
-quarkus.camel.cluster.file.enabled = true
 quarkus.camel.cluster.file.root = target/cluster
 quarkus.camel.cluster.file.acquire-lock-delay = 1s
 quarkus.camel.cluster.file.acquire-lock-interval = 1000
